Rename `WorldBorderSyncerRemover.java`
Since Mixin `class`es are "not `class`es", and therefore will never be referenced in code in a truly object-oriented sense (`extends`, `new`, ...), there is little need to follow established naming conventions here. I will try to name them in the form of an imperative clause, i.e., as commands of a sort, and do away with camel case. Let's see how this helps with readability, among other things.