Add a Mixin to mod `PlayerManager`'s `sendWorldInfo()`
One of the things that a server does after a player collides with a portal block is that it sends information about the common (Overworld) world border (its center, its size, etc.) so that the client can reconstruct it in its newly created `ClientWorld`. This poses a dire problem when we wish to let each dimension's world border be adjusted seperately, however. This Mixin makes the server send the "correct"/corresponding `ServerWorld`'s world border information instead of that of the Overworld.