Remove "environment" fields from fabric.mod.json
They are, for the most part, only confusing, especially in our case--We do want this mod to run on both multiplayer and singleplayer. It was not clear whether "client"/"server" referred to the "logical" ones or the "physical" ones. (It seems that I have missed it being mentioned in HTTPS://FabricMc.net/wiki/Tutorial:Side, but, to be fair, it was one single sentence!)